backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
langchain
langchain-community
qdrant-client
sentence-transformers
httpx
faster-whisper
google-api-python-client
google-auth-httplib2
google-auth-oauthlib


